requirements.txt
streamlit
ultralytics
torch
pillow
opencv-python-headless
numpy
